Use top-level statements
