+++ src/Umbraco.Web.UI/Program.Main.cs
@@ -1,0 +1,21 @@
+using Umbraco.Cms.Web.Common.Hosting;
+
+namespace Umbraco.Cms.Web.UI
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+            => CreateHostBuilder(args)
+                .Build()
+                .Run();
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureUmbracoDefaults()
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStaticWebAssets();
+                    webBuilder.UseStartup<Startup>();
+                });
+    }
+}
