+++ templates/Umbraco.Templates.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <!-- Enable multi-level merging with src -->
   <Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.props', '$(MSBuildThisFileDirectory)../src/'))" />
     
@@ -21,6 +21,10 @@
     <Content Include="UmbracoProject\**" Exclude="bin;obj" />
     <Content Include="..\src\Umbraco.Web.UI\Program.cs">
       <Link>UmbracoProject\Program.cs</Link>
+      <PackagePath>UmbracoProject</PackagePath>
+    </Content>
+    <Content Include="..\src\Umbraco.Web.UI\Program.Main.cs">
+      <Link>UmbracoProject\Program.Main.cs</Link>
       <PackagePath>UmbracoProject</PackagePath>
     </Content>
     <Content Include="..\src\Umbraco.Web.UI\Startup.cs">
